Adding two features consisting of multiple weights
This is sort of a refactoring. It was obviously necessary in order to
support a feature detector of multiple weights. In this case, it is a
feature consisting of one weight per column, meaning that not only is it
more than one, but it can very depending on the board configuration.