Adding --feature-enable-rate as an option.
This value determines the rate or probability of a feature to become
enabled during randomization and mutation. The value defaults to 1 / 6
and the reason for it being lower than half is that there are quite a
lot of features, many with overlapping information detection, and it
doesn't seem to make sense to measure too many at the same time.